Rename SampleCollector to ES5SampleCollector
