+++ src/main/java/io/rathr/audrey/lsp/es5/ES5DocumentService.java
@@ -1,0 +1,4 @@
+package io.rathr.audrey.lsp.es5;
+
+public class ES5DocumentService {
+}
